Core/Misc: implement custom player stat boni
if you want to add an specific amout of any stat to a player you can use the player_bonus_stats table in zyn database.

just set the race and class IDs, specify the level and the boni amount for each stat.

this system is perfect if you don't feel like creating custom items for leveling purposes only