A Ton of Stuff was done...
First of all, added RBAC for some features
Realized my total wrong idea of XP and Rep modification and Changed them to a working state.
Further added a .donator and a .donator port command to limit locations and useabilty of the port command
The Feature with the special rewards seems to be bugged (server crash everytime i killed a mob) since i couldn't figure out whats wrong with it i just commented it out.